refactor: renamed comparison report to regressio
