+++ src/report-templates/index.js
@@ -1,2 +1,2 @@
-export { default as comparisonReportTemplate } from './comparison.js';
+export { default as regressionReportTemplate } from './regression.js';
 export { default as headOnlyReportTemplate } from './head-only.js';
